+++ pysys/launcher/console_run.py
@@ -61,6 +61,7 @@ class ConsoleLaunchHelper(object):
 		self.optionList = ["help","record","purge","verbosity=","type=","trace=","include=","exclude=","cycle=","outdir=",
 			"mode=","modeinclude=","modeexclude=","threads=", "abort=", 'validateOnly', 'vo', 'progress', 'printLogs=', 'grep=', 
 			'ci', 'sort=', 
+			'writer=',
 			]
 
 
@@ -126,8 +127,10 @@ Execution options
 		if printXOptions: printXOptions()
 		print("""
 Advanced:
-   -g, --progress              print progress updates after completion of each test
    -r, --record                use configured 'writers' to record the test results (e.g. XML, JUnit, etc)
+       --writer    CLASSNAME   enables the specified writer (in addition to any writers already enabled automatically), 
+		                       for example "--writer TextResultsWriter" (can be specified multiple times)
+   -g, --progress              enables the writer that prints progress updates after completion of each test
    -p, --purge                 purge files except run.log from the output directory to save space (unless test fails)
    --printLogs     STRING      indicates for which outcome types the run.log output will be printed to the stdout 
                                console; options are: all|none|failures (default is all).
@@ -139,8 +142,8 @@ Advanced:
                                this is a special command for automatically updating the reference files when an 
                                assertDiff fails
 
-The PYSYS_DEFAULT_ARGS environment variable can be used to specify any pysys run arguments that you always wish to use, 
-for example PYSYS_DEFAULT_ARGS=--progress --outdir __pysys_output. 
+The PYSYS_DEFAULT_ARGS environment variable can be used to specify any pysys run arguments that you wish to use on 
+a per-user basis, for example: PYSYS_DEFAULT_ARGS=--progress --writer=TextResultsWriter --outdir=__pysys_output
 
 Selection and filtering options
 -------------------------------
@@ -228,7 +231,11 @@ e.g.
 		# so that it doesn't get enabled with -vDEBUG only -vassertions=DEBUG 
 		# as it is incredibly verbose and slow and not often useful
 		logging.getLogger('pysys.assertions').setLevel(logging.INFO)
-				
+
+		# special hidden dict of extra values to pass to the runner, since we can't change 
+		# the public API now
+		self.userOptions['__extraRunnerOptions'] = __extraRunnerOptions = {}
+
 		for option, value in optlist:
 			if option in ("-h", "--help"):
 				self.printUsage(printXOptions)	  
@@ -321,6 +328,9 @@ e.g.
 				
 			elif option in ["-g", "--progress"]:
 				self.progress = True
+			
+			elif option in ['--writer']:
+				__extraRunnerOptions.setdefault('writerEnable', []).append(value)
 
 			elif option in ["--printLogs"]:
 				printLogs = getattr(PrintLogs, value.upper(), None)
@@ -373,14 +383,12 @@ e.g.
 		# retained for compatibility, but PYSYS_DEFAULT_ARGS is a better way to achieve the same thing
 		if os.getenv('PYSYS_PROGRESS','').lower()=='true': self.progress = True
 		
-		# special hidden dict of extra values to pass to the runner, since we can't change 
-		# the public API now
-		self.userOptions['__extraRunnerOptions'] = {
+		__extraRunnerOptions.update({
 			'progressWritersEnabled':self.progress,
 			'printLogs': printLogs,
 			'printLogsDefault': printLogsDefault, # to use if not provided by a CI writer or cmdline
 			'sort': self.sort
-		}
+		})
 		
 		# load project AFTER we've parsed the arguments, which opens the possibility of using cmd line config in 
 		# project properties if needed